arrumar mais um erro

+++ criarTabelaCartaoCredito.sql
@@ -2,6 +2,6 @@
  idCartao_Credito INT NOT NULL AUTO_INCREMENT,
  Nome_Cartao VARCHAR(45),
  dataValidade_Cartao DATE,
- NumCartao_Cartao INT(16),
+ NumCartao_Cartao float(16),
  Cliente_idCliente INT NOT NULL,
  PRIMARY KEY (idCartao_Credito));
